+++ RUFAS/routines/animal/manure/calf_manure_excretion.py
@@ -1,7 +1,7 @@
 from typing import Tuple, Dict
 
 from RUFAS.general_constants import GeneralConstants
-from ....shared_structures.animal_manure_excretions import AnimalManureExcretions
+from ....data_structures.animal_manure_excretions import AnimalManureExcretions
 from RUFAS.routines.animal.manure.general_manure import (
     calculate_phosphorus_excretion_values,
 )
