+++ RUFAS/routines/animal/manure/dry_cow_manure_excretion.py
@@ -2,7 +2,7 @@ import math
 from typing import Tuple, Dict
 
 from RUFAS.general_constants import GeneralConstants
-from ....shared_structures.animal_manure_excretions import AnimalManureExcretions
+from ....data_structures.animal_manure_excretions import AnimalManureExcretions
 from RUFAS.routines.animal.manure.general_manure import (
     calculate_phosphorus_excretion_values,
 )
