+++ RUFAS/routines/animal/manure/general_manure.py
@@ -1,6 +1,6 @@
 from typing import Tuple
 from RUFAS.general_constants import GeneralConstants
-from ....shared_structures.animal_manure_excretions import AnimalManureExcretions
+from ....data_structures.animal_manure_excretions import AnimalManureExcretions
 
 
 def get_default_animal_manure_excretions() -> AnimalManureExcretions:
