+++ RUFAS/routines/manure/pen_manure/manure_manager_pen.py
@@ -2,8 +2,8 @@ from typing import NamedTuple
 from typing import Set
 
 from RUFAS.routines.manure.pen_manure.pen_manure import PenManure
-from ....shared_structures.animal_combinations import AnimalCombination
-from ....shared_structures.pen_manure_data import PenManureData
+from ....enums import AnimalCombination
+from ....data_structures.pen_manure_data import PenManureData
 
 
 class ManureManagerPen:
