+++ tests/animal_module_tests/test_animal_allocation.py
@@ -9,7 +9,7 @@ from RUFAS.routines.animal.animal_manager import AnimalManager
 from RUFAS.routines.animal.animal_module_constants import AnimalModuleConstants
 from RUFAS.routines.animal.life_cycle.cow import Cow
 from RUFAS.routines.animal.pen import Pen
-from RUFAS.shared_structures.animal_combinations import AnimalCombination
+from RUFAS.enums import AnimalCombination
 
 
 def test_get_dry_cows(mocker: MockerFixture) -> None:
