+++ tests/animal_module_tests/test_animal_grouping_scenarios.py
@@ -8,7 +8,7 @@ from RUFAS.routines.animal.life_cycle.calf import Calf
 from RUFAS.routines.animal.life_cycle.heiferI import HeiferI
 from RUFAS.routines.animal.life_cycle.heiferII import HeiferII
 from RUFAS.routines.animal.life_cycle.heiferIII import HeiferIII
-from RUFAS.shared_structures.animal_combinations import AnimalCombination
+from RUFAS.enums import AnimalCombination
 
 
 def test_constructor() -> None:
