+++ tests/animal_module_tests/test_pen.py
@@ -13,10 +13,10 @@ from RUFAS.routines.animal.life_cycle.cow import Cow
 from RUFAS.routines.animal.life_cycle.heiferI import HeiferI
 from RUFAS.routines.animal.life_cycle.heiferII import HeiferII
 from RUFAS.routines.animal.life_cycle.heiferIII import HeiferIII
-from RUFAS.shared_structures.animal_manure_excretions import AnimalManureExcretions
+from RUFAS.data_structures.animal_manure_excretions import AnimalManureExcretions
 from RUFAS.routines.animal.pen import Pen
-from RUFAS.shared_structures.animal_combinations import AnimalCombination
-from RUFAS.shared_structures.pen_manure_data import PenManureData
+from RUFAS.enums import AnimalCombination
+from RUFAS.data_structures.pen_manure_data import PenManureData
 
 
 @pytest.fixture
