+++ tests/test_manure_module/test_animal_manure_excretions/test_lactating_cow_manure_excretion.py
@@ -5,7 +5,7 @@ from pytest import approx
 from pytest_mock import MockerFixture
 
 from RUFAS.general_constants import GeneralConstants
-from RUFAS.shared_structures.animal_manure_excretions import AnimalManureExcretions
+from RUFAS.data_structures.animal_manure_excretions import AnimalManureExcretions
 from RUFAS.routines.animal.manure.lactating_cow_manure_excretion import (
     manure_calculations,
 )
