+++ tests/test_manure_module/test_manure_manager_pen.py
@@ -5,8 +5,8 @@ from RUFAS.routines.animal.life_cycle.cow import Cow
 from RUFAS.routines.animal.pen import Pen
 from RUFAS.routines.manure.pen_manure.pen_manure import PenManure
 from RUFAS.routines.manure.pen_manure.manure_manager_pen import ManureManagerPen
-from RUFAS.shared_structures.animal_combinations import AnimalCombination
-from RUFAS.shared_structures.pen_manure_data import PenManureData
+from RUFAS.enums import AnimalCombination
+from RUFAS.data_structures.pen_manure_data import PenManureData
 
 
 def test_manure_manager_pen_init(mocker: MockerFixture) -> None:
