+++ tests/test_manure_module/test_pen_manure.py
@@ -3,7 +3,7 @@ from dataclasses import fields
 from pytest import approx
 
 from RUFAS.general_constants import GeneralConstants
-from RUFAS.shared_structures.animal_manure_excretions import AnimalManureExcretions
+from RUFAS.data_structures.animal_manure_excretions import AnimalManureExcretions
 from RUFAS.routines.manure.constants_and_units.manure_constants import ManureConstants
 from RUFAS.routines.manure.pen_manure.pen_manure import PenManure
