+++ tests/test_simulation_engine.py
@@ -5,7 +5,7 @@ from pytest_mock import MockerFixture
 from RUFAS.routines import Feed
 from RUFAS.routines.EEE.EEE_manager import EEEManager
 from RUFAS.simulation_engine import SimulationEngine
-from RUFAS.shared_structures.pen_manure_data import PenManureData
+from RUFAS.data_structures.pen_manure_data import PenManureData
 from RUFAS.time import Time
